requirements.txt
pillow
matplotlib
requests
ImageHash
flask
plyer
matplotlib
